Move webhook to a standalone package under modules
